SAM-6D/requirements.txt
torch
torchvision
fvcore
xformers
torchmetrics
blenderproc
opencv-python
omegaconf
ruamel.yaml
hydra-colorlog
hydra-core
gdown
pandas
imageio
pyrender
pytorch-lightning
pycocotools
distinctipy
git+https://github.com/facebookresearch/segment-anything.git
git+https://github.com/facebookresearch/sam2
ultralytics
timm
gorilla-core
trimesh
gpustat
imgaug
einops
